services/gateway/requirements.txt
fastapi
pydantic
pydantic_core
python-dotenv
uvicorn
pika
httpx
PyJWT
requests
python-multipart
python-jose
passlib
requests
pymysql
cryptography
bcrypt
sqlalchemy
prometheus-client